app/requirements.txt
grpcio==1.63.0
grpcio_reflection==1.63.0
protobuf==5.26.1
psycopg[binary]==3.1.19
